+++ CHANGELOG.md
@@ -22,6 +22,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   CNBs. ([#59](https://github.com/heroku/buildpacks-deb-packages/pull/59))
 - Modified the buildpack output format to align with our other
   CNBs. ([#60](https://github.com/heroku/buildpacks-deb-packages/pull/60))
+- `PackageNotFound` errors now return a list of suggested package names.
+  ([#69](https://github.com/heroku/buildpacks-deb-packages/pull/69))
 
 ## [0.0.1] - 2024-10-10